basic: fileio: fallback to EIO if no errno is set by fputs
fputs isn's supposed to set errno at all, this is a glibc'ism, we
are supposed to call ferror and feof to figure out what happend.

This commit reuses glibc's errno and sets EIO on musl, as musl is
implementing this more correctly, while staying forward compatible
with glibc if it decides to fix this. ferror explicitly says that
no errno should be set with it while fputs doesn't mention errno
at all.